+++ tests/testthat/test-multi-req.R
@@ -21,6 +21,14 @@ test_that("requests happen in parallel", {
   expect_lt(time[[3]], 1)
 })
 
+test_that("can perform >128 file uploads in parallel", {
+  temp <- withr::local_tempfile(lines = letters)
+  req <- request(example_url()) %>% req_body_file(temp)
+  reqs <- rep(list(req), 150)
+  
+  expect_no_error(req_perform_parallel(reqs, on_error = "continue"))
+})
+
 test_that("can download files", {
   reqs <- list(request_test("/json"), request_test("/html"))
   paths <- c(withr::local_tempfile(), withr::local_tempfile())