+++ NEWS.md
@@ -1,5 +1,7 @@
 # httr2 (development version)
 
+* Corrupt `rds` files no longer cause the request to fail.
+
 # httr2 1.0.1
 
 * `req_perform_stream()` gains a `round = c("byte", "line")` argument to control