+++ NEWS.md
@@ -1,6 +1,7 @@
 # httr2 (development version)
 
 * `req_url_query()` now allows you to opt out of escaping for multi-value parameters (#404).
+* `req_perform_parallel()` now works correctly with `req_cache()` (#447).
 * `req_body_file()` now only opens a connection when the request actually needs data. In particular, this makes it work better with `req_perform_parallel()` (#487).
 * `req_perform_parallel()` now works when downloading 0 byte files (#478)
 * Corrupt `rds` files no longer cause the request to fail.