+++ NEWS.md
@@ -2,9 +2,13 @@
 
 * `req_perform_stream()` no longer applies the `callback` to unsuccessful
   responses, instead creating a regular response. It also now sets `last_request()` and `last_response()` (#479).
+* `req_cache()` now defaults the `debug` argument to the `httr2_cache_debug` option to make it easier to debug caching buried in other people's code (#486).
+* `req_oauth_password()` now only asks for your password once (#498).
+* `req_url_query()` now allows you to opt out of escaping for multi-value parameters (#404).
+* `req_perform_parallel()` now works correctly with `req_cache()` (#447).
 * `req_body_file()` now only opens a connection when the request actually needs data. In particular, this makes it work better with `req_perform_parallel()` (#487).
 * `req_perform_parallel()` now works when downloading 0 byte files (#478)
-* Corrupt `rds` files no longer cause the request to fail.
+* `req_cache()` no longer fails if the `rds` files are somehow corrupted
 
 # httr2 1.0.1